requirements.txt
Flask
tensorflow
numpy
Pillow
h5py
